requirements.txt
PyYAML￼
numpy 
mistralai